Convert node evacuation to multi-group
This patch does the necessary changes to make the new tryMGEvac work
correctly: each instance remains inside its primary node's group when
it is evacuated.

This is done by splitting up the to-be-evacuated instance list per
group, together with the group node and instance list, and then
running the old tryEvac for each group.

Signed-off-by: Iustin Pop <iustin@google.com>
Reviewed-by: Balazs Lecz <leczb@google.com>