Fix premature abort of LUOobCommand due to result.Raise
This is a bug I recognized while doing tests on gnt-node health. A left
over result.Raise line causes premature abort of LUOobCommand on the
first node failing the RPC call. This is not expected behaviour for
multiple nodes.

This bug is not critical for the current use case as for a single node
this just works as expected. Though for multi node this will fail,
that's the reason why QA and tests passed as currently we just had
single node invocation. QA for multi node will be included in the
gnt-node health, as this is the first command using multi node.

Signed-off-by: René Nussbaumer <rn@google.com>
Reviewed-by: Iustin Pop <iustin@google.com>